Some fixes on file organization
