+++ app/tasks/exceptions.py
@@ -1,0 +1,11 @@
+class SchedulerTasksException(Exception):
+    pass
+
+
+class SchedulerHTTPException(SchedulerTasksException):
+    def __init__(self, value, desc=None):
+        self.value = value
+        self.desc = desc
+
+    def __str__(self):
+        return repr(self.value)
