+++ app/tasks/tasks.py
@@ -8,18 +8,7 @@ from datetime import timedelta
 from sqlalchemy.orm import sessionmaker
 from utilities.request2 import Request2
 
-
-class SchedulerTasksException(Exception):
-    pass
-
-
-class SchedulerHTTPException(SchedulerTasksException):
-    def __init__(self, value, desc=None):
-        self.value = value
-        self.desc = desc
-
-    def __str__(self):
-        return repr(self.value)
+from exceptions import SchedulerHTTPException
 
 
 class TaskLogic:
